Remove outdated dependency, add rails
