+++ table_helper.gemspec
@@ -14,7 +14,8 @@ Gem::Specification.new do |s|
   s.test_files        = `git ls-files -- test/*`.split("\n")
   s.rdoc_options      = %w(--line-numbers --inline-source --title table_helper --main README.rdoc)
   s.extra_rdoc_files  = %w(README.rdoc CHANGELOG.rdoc LICENSE)
-  
+
+  s.add_runtime_dependency('rails')
   s.add_development_dependency("rake")
-  s.add_development_dependency("plugin_test_helper", ">= 0.3.2")
+  s.add_development_dependency("sqlite3")
 end
