--- test/app_root/app/models/person.rb
@@ -1,2 +1,0 @@
-class Person < ActiveRecord::Base
-end
